feat: add validatos and basic exception handling

+++ TD.LeaveManaged.Application/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocationsCommandHandler.cs
@@ -5,6 +5,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using TD.LeaveManaged.Application.DTOs.LeaveAllocation.Validators;
+using TD.LeaveManaged.Application.Exceptions;
 using TD.LeaveManaged.Application.Features.LeaveAllocations.Requests.Commands;
 using TD.LeaveManaged.Application.Persistence.Contracts;
 
@@ -13,16 +15,24 @@ namespace TD.LeaveManaged.Application.Features.LeaveAllocations.Handlers.Command
     public class DeleteLeaveAllocationsCommandHandler : IRequestHandler<DeleteLeaveAllocationsCommand>
     {
         private readonly ILeaveAllocationRepository _leaveAllocationRepository;
-        private readonly IMapper _mapper;
+        
 
-        public DeleteLeaveAllocationsCommandHandler(ILeaveAllocationRepository leaveAllocationRepository, IMapper mapper)
+        public DeleteLeaveAllocationsCommandHandler(ILeaveAllocationRepository leaveAllocationRepository)
         {
-            this._leaveAllocationRepository = leaveAllocationRepository;
-            this._mapper = mapper;
+            _leaveAllocationRepository = leaveAllocationRepository;            
         }
         public async Task<Unit> Handle(DeleteLeaveAllocationsCommand request, CancellationToken cancellationToken)
         {
+            var validator = new DeleteLeaveAllocationDtoValidator();
+            var validationResult = await validator.ValidateAsync(request.Id);
+            if (validationResult.IsValid == false)
+                throw new ValidationException(validationResult);
+
+                        
             var leaveAllocation = await _leaveAllocationRepository.Get(request.Id);
+            if (leaveAllocation == null)
+                throw new NotFoundException(nameof(leaveAllocation), request.Id);
+
             await _leaveAllocationRepository.Delete(leaveAllocation);
             return Unit.Value;
         }
